Parser.java: Reduce code indentation
convertTextToTask method has more than 3 levels of indentation,
which makes the code harder to read and maintain.

The deeper the nesting, the harder it is for the reader to keep track
of the logic. By reducing the number of indentation levels, the logic
becomes clearer and easy to follow.

Refactor the method by initializing a variable and move the try catch
to the higher level improve error handling without additional nesting.
Create a new loadIdDone method to keeps the switch block concise
and focused only on task creation.

Let's decrease the number of levels of indentation to increase code
readability and clarity.