Storage.java: Restructure code and throw exception
Restructure the code in loadFileContent method to simplifies the flow
and focuses on processing the lines.

In loadFileContent method, the FileNotFoundException is handled right
at the point where it can be thrown (new Scanner(file)), making it
easier to trace the cause of the error.

Throw exception to signal that there might be an unexpected error
occurred during the program execution.

The code that called the function can then "catch" this exception
using a try-catch block, allowing it to take appropriate actions

Let's
- throws exception for other method to handle
- restructure the code to increase readability