TaskList.java: Revise code
Long method often contains more information than what the reader can
process at a time. Shortening the code to smoothen the flow and
increase the quality of code.

Avoid repetitive code to reduce the LoC of addTask method and increase
the readability of the code.

Extract a new method to reduce the level of indentation in findTask
method. The for loop in findTask no longer has nested loops, making it
easier to read. Each method does one thing, improving clarity and
maintainability.

Let's
- avoid long method
- reduce the indentation level of code